% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFirstOrderMarginals}
\alias{getFirstOrderMarginals}
\title{Function to get the first order marginals}
\usage{
getFirstOrderMarginals(permutations, smoothed = FALSE)
}
\arguments{
\item{permutations}{a list of objects of class \code{\linkS4class{Permutation}}}

\item{smoothed}{a logical value to indicate whether the estimation should
be smoothed (\code{FALSE} by default)}
}
\value{
A matrix containing the first order marginals
}
\description{
This function estimates, from a list of permutations, the first
order marginals
}
\examples{
scores <- matrix(rnorm(100, 1, 3), ncol=10)
permus <- getPermutationFromScores (scores)
getFirstOrderMarginals(permus)

}
