% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getIdentityPermutation}
\alias{getIdentityPermutation}
\title{Function to generate an identity permutation}
\usage{
getIdentityPermutation(size)
}
\arguments{
\item{size}{numeric value indicating the size of the permutation}
}
\value{
A new object of class \code{\linkS4class{Permutation}} equal to the
identity permutation of size \code{size}
}
\description{
This function creates an identity permutation
}
