% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPermutationFromScores}
\alias{getPermutationFromScores}
\title{Function to generate permutations from scores}
\usage{
getPermutationFromScores(scores, type = "ranking", decreasing = FALSE)
}
\arguments{
\item{scores}{either a vector or a matrix with the scores. Each set of
scores has to be a row in the matrix}

\item{type}{a character argument indicating the type of permutation.
Valid values are \code{'ranking'} (default value) and \code{'ordering'}}

\item{decreasing}{a logical value to indicate which score is better. If
\code{FALSE} (default value), higher scores are better}
}
\value{
Either an object or a list of objects of class
\code{\linkS4class{Permutation}} corresponding to the permutations derived
from the scores
}
\description{
This function creates a permutation or a set of permutations
based on one or more sets of scores or ratings
}
\examples{
scores <- matrix(rnorm(100, 1, 3), ncol=10)
permus <- getPermutationFromScores (scores)
scores[5, ]
permus[[5]]

}
