% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRandomPermutation}
\alias{getRandomPermutation}
\title{Function to generate a random permutation}
\usage{
getRandomPermutation(size)
}
\arguments{
\item{size}{numeric value indicating the size of the permutation}
}
\value{
A new object of class \code{\linkS4class{Permutation}} representing
a random permutation of size \code{size}
}
\description{
This function creates a random permutation
}
