% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation.R
\name{\%.\%,Permutation-method}
\alias{\%.\%,Permutation-method}
\title{Operator to compose two permutations}
\usage{
\S4method{\%.\%}{Permutation}(e1, e2)
}
\arguments{
\item{e1}{Object of class \code{\linkS4class{Permutation}} representing the
firs permutation in the composition}

\item{e2}{Object of class \code{\linkS4class{Permutation}} representing the
second permutation in the composition}
}
\value{
A new object of class \code{\linkS4class{Permutation}} equal to the
composition of e1 and e2
}
\description{
This operator returns the composition of two permutations
}
\examples{
permu <- permutation(order(runif(10)))
permu \%.\% inverse(permu)
}
