% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation.R
\name{inverse,Permutation-method}
\alias{inverse,Permutation-method}
\title{Function to invert a permutation}
\usage{
\S4method{inverse}{Permutation}(x)
}
\arguments{
\item{permutation}{Object of class \code{\linkS4class{Permutation}} to be
inverted}
}
\value{
A new object of class \code{\linkS4class{Permutation}} equal
to the inverse of \code{permutation}
}
\description{
This returns the inverse of a permutation
}
\examples{
permu <- permutation(order(runif(10)))
permu
inverse(permu)

}
