% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation.R
\docType{class}
\name{Permutation-class}
\alias{Permutation-class}
\title{An S4 class to represent Permutations}
\description{
An S4 class to represent Permutations
}
\section{Slots}{

\describe{
\item{\code{permutation}}{Numeric vector containing the actual permutation}

\item{\code{size}}{Numeric value indicating the size of the permutation}
}}

