% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation.R
\name{permutation}
\alias{permutation}
\title{Basic constructor of the Permutation class}
\usage{
permutation(vector)
}
\arguments{
\item{vector}{A numeric vector containing the permutation}
}
\value{
An object of class \code{\linkS4class{Permutation}} that represents
the permutation passes as an argument
}
\description{
This function creates an object of class \code{\linkS4class{Permutation}}
}
